requirements.txt
beautifulsoup4==4.10.0
Pillow==8.4.0
requests==2.25.1
